Cleaning
Cleaned up code in 464 server files

+++ Server/wServer/logic/behaviors/Order.cs
@@ -38,7 +38,7 @@ namespace wServer.logic.behaviors
         protected override void TickCore(Entity host, RealmTime time, ref object state)
         {
             if (targetState == null)
-                targetState = FindState(host.Manager.Behaviors.Definitions[(ushort) children].Item1, targetStateName);
+                targetState = FindState(host.Manager.Behaviors.Definitions[children].Item1, targetStateName);
             foreach (Entity i in host.GetNearestEntities(range, children))
                 if (!i.CurrentState.Is(targetState))
                     i.SwitchTo(targetState);
